python/tools/camera_stats/server/requirement.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mock==1.0.1
pymongo==3.2.1
PyYAML==3.13
requests==2.21.0
schematics==1.1.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
xlrd==1.2.0
xlwt==1.3.0
threadpool==1.3.2
honcho==1.0.1
gunicorn==19.9.0